# src/main_service/api/api.yaml
openapi: 3.0.0
info: 
  title: Main Servide API
  version: 1.0.0
paths:
  /signup:
    post:
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Успешная регистрация
          headers:
            Set-cookie:
              description: Куки для работы с сервисом
              schema:
                type: string
        '403':
          description: Пользователь уже существует
  /login:
    post:
      summary: Аутентификация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Успешная авторизация
          headers:
            Set-cookie:
              description: Куки для работы с сервисом
              schema:
                type: string
        '401':
          description: Пользователя с таким логином не существует
        '403':
          description: Неверный пароль к логину
  /update:
    post:
      security:
        - AuthToken: []
      summary: Обновление данных пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  nullable: true
                lastName:
                  type: string
                  nullable: true
                birthDate:
                  type: string
                  format: date
                  nullable: true
                mail:
                  type: string
                  format: email
                  nullable: true
                phoneNumber:
                  type: string
                  nullable: true        
      responses:
        '200':
          description: Успешное обновление данных пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  firstName:
                    type: string
                    nullable: true
                  lastName:
                    type: string
                    nullable: true
                  birthDate:
                    type: string
                    format: date
                    nullable: true
                  mail:
                    type: string
                    format: email
                    nullable: true
                  phoneNumber:
                    type: string
                    nullable: true
        '400': 
          description: Нет куки
        '401':
          description: Невалидная кука
  /tasks/create:
    post:
      security:
        - AuthToken: []
      summary: Создание задачи
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
                status:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Успешное создание задачи
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  content:
                    type: string
                  status:
                    type: string
        '400': 
          description: Нет куки
        '401':
          description: Невалидная кука
        '403':
          description: В запросе нет title или content
  /tasks/update:
    post:
      security:
        - AuthToken: []
      summary: Обновление задачи
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                taskId:
                  type: integer
                title:
                  type: string
                  nullable: true
                content:
                  type: string
                  nullable: true
                status:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Успешное обновление задачи
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  content:
                    type: string
                  status:
                    type: string
        '400': 
          description: Нет куки
        '401':
          description: Невалидная кука
        '403':
          description: В запросе нет taskId
        '404':
          description: Невалидный запрос
        '405':
          description: Невалидный taskId
  /tasks/delete:
    post:
      security:
        - AuthToken: []
      summary: Удаление задачи
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                taskId:
                  type: integer
      responses:
        '200':
          description: Успешное удаление задачи
        '400': 
          description: Нет куки
        '401':
          description: Невалидная кука
        '403':
          description: В запросе нет taskId
        '405':
          description: Невалидный taskId
  /tasks/get:
    post:
      security:
        - AuthToken: []
      summary: Получить задачу
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                taskId:
                  type: integer
      responses:
        '200':
          description: Успешное получение задачи
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  content:
                    type: string
                  status:
                    type: string
        '400': 
          description: Нет куки
        '401':
          description: Невалидная кука
        '403':
          description: В запросе нет taskId
        '405':
          description: Невалидный taskId
  /tasks/get_page:
    post:
      security:
        - AuthToken: []
      summary: Получить страницу с задачами
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pageSize:
                  type: integer
                page:
                  type: integer
      responses:
        '200':
          description: Успешное получение страницы с задачами
          content:
            application/json:
              schema:
                type: array
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  content:
                    type: string
                  status:
                    type: string
        '400': 
          description: Нет куки
        '401':
          description: Невалидная кука
        '403':
          description: В запросе нет pageSize или page\
components:
  securitySchemes:
    AuthToken:
      type: http
      scheme: bearer
      bearerFormat: JWT